Production 46/backendNotificationEmailService/pyproject.toml
[project]
name = "backend2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=4.0.1",
    "alembic>=1.15.2",
    "fastapi[uvicorn]>=0.115.12",
    "psycopg>=3.2.7",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.4",
    "python-jose[cryptography]>=3.4.0",
    "pyyaml>=6.0.2",
    "sqlalchemy-utils>=0.41.2",
    "uvicorn>=0.34.2",
]
